+++ tests/integration/service/test_feature_table_cache.py
@@ -13,76 +13,10 @@ from featurebyte.query_graph.sql.common import sql_to_string
 from tests.util.helper import create_observation_table_from_dataframe
 
 
-@pytest.fixture(name="event_view")
-def event_view_fixture(event_table):
-    # create event view
-    event_view = event_table.get_view()
-    assert event_view.columns == [
-        "ËVENT_TIMESTAMP",
-        "CREATED_AT",
-        "CUST_ID",
-        "ÜSER ID",
-        "PRODUCT_ACTION",
-        "SESSION_ID",
-        "ÀMOUNT",
-        "TZ_OFFSET",
-        "TRANSACTION_ID",
-        "EMBEDDING_ARRAY",
-        "ARRAY",
-        "FLAT_DICT",
-        "NESTED_DICT",
-    ]
-    return event_view
-
-
-@pytest.fixture(name="feature_group")
-def feature_group_fixture(event_view):
-    """
-    Fixture for a simple FeatureGroup with count features
-    """
-    event_view["derived_value_column"] = 1.0 * event_view["ÜSER ID"]
-    feature_group = event_view.groupby("ÜSER ID").aggregate_over(
-        method="count",
-        windows=["2h", "24h"],
-        feature_names=["COUNT_2h", "COUNT_24h"],
-    )
-    feature_group["COUNT_2h / COUNT_24h"] = feature_group["COUNT_2h"] / feature_group["COUNT_24h"]
-    return feature_group
-
-
-@pytest.fixture(name="feature_group_per_category")
-def feature_group_per_category_fixture(event_view):
-    """
-    Fixture for a FeatureGroup with dictionary features
-    """
-
-    feature_group_per_category = event_view.groupby(
-        "ÜSER ID", category="PRODUCT_ACTION"
-    ).aggregate_over(
-        method="count",
-        windows=["2h", "24h"],
-        feature_names=["COUNT_BY_ACTION_2h", "COUNT_BY_ACTION_24h"],
-    )
-    # add features based on transformations on count per category
-    feature_counts_24h = feature_group_per_category["COUNT_BY_ACTION_24h"]
-    feature_group_per_category["ENTROPY_BY_ACTION_24h"] = feature_counts_24h.cd.entropy()
-    feature_group_per_category["MOST_FREQUENT_ACTION_24h"] = feature_counts_24h.cd.most_frequent()
-    feature_group_per_category["NUM_UNIQUE_ACTION_24h"] = feature_counts_24h.cd.unique_count()
-    feature_group_per_category[
-        "NUM_UNIQUE_ACTION_24h_exclude_missing"
-    ] = feature_counts_24h.cd.unique_count(include_missing=False)
-
-    feature_counts_2h = feature_group_per_category["COUNT_BY_ACTION_2h"]
-    feature_group_per_category[
-        "ACTION_SIMILARITY_2h_to_24h"
-    ] = feature_counts_2h.cd.cosine_similarity(feature_counts_24h)
-
-    return feature_group_per_category
-
-
 @pytest.fixture(name="feature_list")
 def feature_list_fixture(feature_group, feature_group_per_category):
     """Feature List fixture"""
+    feature_group["COUNT_2h / COUNT_24h"] = feature_group["COUNT_2h"] / feature_group["COUNT_24h"]
     feature_list = FeatureList(
         [
             feature_group["COUNT_2h"],
@@ -103,6 +37,7 @@ def feature_list_fixture(feature_group, feature_group_per_category):
 @pytest.fixture(name="two_feature_lists")
 def two_feature_lists_fixure(feature_group, feature_group_per_category):
     """Two Feature Lists fixture"""
+    feature_group["COUNT_2h / COUNT_24h"] = feature_group["COUNT_2h"] / feature_group["COUNT_24h"]
     feature_list_1 = FeatureList(
         [
             feature_group["COUNT_2h"],
